Nk2303/6 20 add icon fav box to board and update bg (#46)
* css for boards

* boards css done

* boards-css-90%-complete

* resolve conflict again

* profile css in progress

* handle create without input

* background for main

* css for background in progress"

* add icon favorite box to board and update bg
