+++ src/container/Home.js
@@ -17,8 +17,7 @@ const Home = ({userInfo}) => {
                     <Col xs={12} md={9}>
                         <h1 className="display-3">Welcome to Plany Boat!</h1>
                         <br/>
-                        { userInfo ? 
-                        <p className="text-info">Hi, {userInfo.username} !</p>
+                        { userInfo ? <p className="text-info">Hi, {userInfo.username} !</p>
                         :
                         <Row> 
                             <Col xs={12} md={5}>
@@ -43,7 +42,7 @@ const Home = ({userInfo}) => {
                     </Col>
                 </Row>
                 <hr className="my-4"></hr>
-                <p className="size22">This is a complex app, created by two programing sailors: Kim and David.</p>
+                <p className="size22">This is a complex and interactive app, created by two programing sailors: Kim and David.</p>
                 <p className="size22"></p>
             </Container>
