+++ src/index.css
@@ -1,10 +1,7 @@
 
 body {
   margin: 0;
- 
- 
-    
-    
+
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
